Implement read-only transactions
Utilise read-only transactions for applicable queries
Utilise read-only transactions for all transactions on virtual collections
Handle query for "id" field within a reference
